Moving examples to a separate class
